requirements.txt
google-auth-oauthlib
google-api-python-client
openai
python-dotenv
protobuf
groq
cryptography